add DeleteFamilyVersion test case and pass
